+++ Tests/Git2SwiftTests/TagTest.swift
@@ -8,6 +8,10 @@
 
 import XCTest
 
+import Foundation
+
+@testable import Git2Swift
+
 class TagTest: XCTestCase {
     
     static let name = "tagTest"